feat: add dto validation, code refactoring
