--- src/User/types.ts
@@ -1,18 +1,0 @@
-export interface IUser {
-  id: string;
-  login: string;
-  password: string;
-  version: number;
-  createdAt: number;
-  updatedAt: number;
-}
-
-export interface CreateUserDto {
-  login: string;
-  password: string;
-}
-
-export interface UpdatePasswordDto {
-  oldPassword: string;
-  newPassword: string;
-}
