+++ src/User/user.controller.ts
@@ -1,5 +1,14 @@
-import { Controller, Get, HttpCode, Param, Post, Req } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Get,
+  HttpCode,
+  Param,
+  Post,
+  // Req,
+} from '@nestjs/common';
 import { UserService } from './user.service';
+import { CreateUserDto, FindID } from './Dto/types';
 
 @Controller('user')
 export class UserController {
@@ -10,12 +19,14 @@ export class UserController {
     return this.userService.getAllUsers();
   }
   @Get(':id')
-  getUserById(@Param('id') id: string): string {
-    return this.userService.getUserById(id);
+  getUserById(@Param() params: FindID): string {
+    return this.userService.getUserById(params);
   }
   @Post()
   @HttpCode(201)
-  create(@Req() req: Request): string {
-    return this.userService.create(req);
+  create(@Body() createUserDto: CreateUserDto): string {
+    // console.log(createUserDto);
+    // return 'Create user';
+    return this.userService.create(createUserDto);
   }
 }
