+++ src/User/user.service.ts
@@ -1,16 +1,17 @@
 import { Injectable } from '@nestjs/common';
+import { CreateUserDto, FindID } from './Dto/types';
 
 @Injectable()
 export class UserService {
   getAllUsers(): string {
     return 'getAllUsers';
   }
-  getUserById(id: string): string {
+  getUserById(id: FindID): string {
     console.log(id);
     return 'getUserById';
   }
-  create(req: Request): string {
-    console.log(req.body);
+  create(user: CreateUserDto): string {
+    console.log(user);
     return 'createUser';
   }
 }
